Revert "Minor clean up"
This reverts commit 428b187722a8cc72872a30d7eed3eff67043e691.

+++ src/main/java/resonantinduction/electrical/battery/TileBattery.java
@@ -137,7 +137,7 @@ public class TileBattery extends TileElectrical implements IConnector<BatteryNet
 	{
 		ArrayList data = new ArrayList();
 		data.add(energy.getEnergy());
-		data.add(ioMap);
+		data.add((short) ioMap);
 		return data;
 	}
